+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.5.31 Release:
+## Update Notes:
+- This is not a V14 update. An update for V14 is still a work in progress.
+## Bug Fixes:
+- Channel Divinity (Cleric, modern)
+- Arcane Ward
+- Tinker's Magic
+## API Update Notes
+- `CompendiumBrowser.select` has new filters, `compendium` and `systemIdentifier`.
+
 # 1.5.30 Release:
 ## Update Notes:
 - This is not a V14 update. An update for V14 is still a work in progress.